requirements.txt
python-telegram-bot
playwright
asyncio
